backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
pydantic
pydantic-settings
python-dotenv
transformers
torch
accelerate